More sensible way to prepare peaks for sinogram
